+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.5.0 - 2023-02-01
+- Added tooltip override via `extras.tooltip`
+
+## 2.4.2 - 2023-02-01
+- Fixed hover title on column chart
+
 ## 2.4.1 - 2022-02-18
 - Fixed xAxis auto-categories for MultiColumnChart